fix some circular imports
