+++ theano/gof/compiledir.py
@@ -1,58 +1,19 @@
 from __future__ import print_function
 import six.moves.cPickle as pickle
-import errno
 import logging
 import os
-import platform
-import re
 import shutil
-import struct
-import socket
-import sys
 
 import numpy
 
 import theano
 from six import string_types, iteritems
-from theano.configparser import config, AddConfigVar, ConfigParam, StrParam
+from theano.configparser import config
 from theano.gof.utils import flatten
 from theano.misc.windows import output_subprocess_Popen
 
 
 _logger = logging.getLogger("theano.gof.compiledir")
-
-try:
-    p_out = output_subprocess_Popen([theano.config.cxx, '-dumpversion'])
-    gcc_version_str = p_out[0].strip().decode()
-except OSError:
-    # Typically means gcc cannot be found.
-    gcc_version_str = 'GCC_NOT_FOUND'
-
-
-def local_bitwidth():
-    """
-    Return 32 for 32bit arch, 64 for 64bit arch.
-
-    By "architecture", we mean the size of memory pointers (size_t in C),
-    *not* the size of long int, as it can be different.
-
-    """
-    # Note that according to Python documentation, `platform.architecture()` is
-    # not reliable on OS X with universal binaries.
-    # Also, sys.maxsize does not exist in Python < 2.6.
-    # 'P' denotes a void*, and the size is expressed in bytes.
-    return struct.calcsize('P') * 8
-
-
-def python_int_bitwidth():
-    """
-    Return the bit width of Python int (C long int).
-
-    Note that it can be different from the size of a memory pointer.
-
-    """
-    # 'l' denotes a C long int, and the size is expressed in bytes.
-    return struct.calcsize('l') * 8
 
 
 def cleanup():
